move RunnerContext to own file
